requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
Django==5.1.1
Flask==3.0.3
fonttools==4.54.1
image==1.5.33
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
Werkzeug==3.0.4
